+++ .gitignore
@@ -1,7 +1,7 @@
 logs/*
 !.gitkeep
 node_modules/
-app/lib/
 tmp
 .DS_Store
 .idea
+dist/
