+++ README.md
@@ -12,18 +12,18 @@ This project is building by me to practice AngularJS (version 1.7.5 now).
 
 ## Features
 
-&check; Add notes&NewLine;
-&check; Delete notes&NewLine;
+&check; Add notes\
+&check; Delete notes
 
 ### Prerequisites
 
-&Tilde; Edit notes by double-clicking (add double-tap for mobile)&NewLine;
-&minus; Add empty list template, about page and 404 page&NewLine;
-&Tilde; Add stylesheet&NewLine;
-&minus; Add creation date to note and sort it by oldest and newest&NewLine;
-&minus; Check/uncheck notes&NewLine;
-&times; Filters for complete/active notes&NewLine;
-&minus; Filter notes by text&NewLine;
-&minus; Counts how many notes are left&NewLine;
-&Tilde; Notes are saved to localStorage (move local storage functions to external service)&NewLine;
-&minus; Pagination list&NewLine;
+&Tilde; Edit notes by double-clicking (add double-tap for mobile)\
+&minus; Add empty list template, about page and 404 page\
+&Tilde; Add stylesheet\
+&minus; Add creation date to note and sort it by oldest and newest\
+&minus; Check/uncheck notes\
+&times; <del>Filters for complete/active notes</del>\
+&minus; Filter notes by text\
+&minus; Counts how many notes are left\
+&Tilde; Notes are saved to localStorage (move local storage functions to external service)\
+&minus; Pagination list
