+++ app/app.css
@@ -1,7 +1,3 @@
-@import "/header/header.css";
-@import "/note-form/note-form.css";
-@import "/note-list/note-list.css";
-
 * {
   margin: 0;
   padding: 0;
