+++ app/index.html
@@ -14,23 +14,9 @@
       href="https://fonts.googleapis.com/css?family=Lato&display=swap"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="app.css" />
-    <link rel="shortcut icon" href="snn-icon.png" type="image/x-icon" />
   </head>
   <body ng-controller="appCtrl">
     <header></header>
     <div ng-view></div>
-    <!-- In production use:
-  <script src="//ajax.googleapis.com/ajax/libs/angularjs/x.x.x/angular.min.js"></script>
-  -->
-    <script src="lib/angular/angular.js"></script>
-    <script src="lib/angular-route/angular-route.js"></script>
-    <script src="app.module.js"></script>
-    <script src="note-form/note-form.module.js"></script>
-    <script src="note-form/note-form.component.js"></script>
-    <script src="header/header.module.js"></script>
-    <script src="header/header.component.js"></script>
-    <script src="note-list/note-list.module.js"></script>
-    <script src="note-list/note-list.component.js"></script>
   </body>
 </html>
