+++ package.json
@@ -4,29 +4,34 @@
   "version": "0.0.0",
   "description": "This is a to-do-list type app that manages a bunch of short notes.",
   "dependencies": {
-    "angular": "^1.7.5",
+    "angular": "^1.7.8",
     "angular-loader": "^1.7.5",
-    "angular-route": "^1.7.5"
+    "angular-route": "^1.7.8"
   },
   "devDependencies": {
+    "@babel/core": "^7.7.2",
+    "@babel/preset-env": "^7.7.1",
     "angular-mocks": "^1.7.5",
-    "cpx": "^1.5.0",
-    "http-server": "^0.11.1",
+    "babel-loader": "^8.0.6",
+    "css-loader": "^3.2.0",
+    "file-loader": "^4.2.0",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
     "jasmine-core": "^3.3.0",
     "karma": "^4.4.1",
     "karma-chrome-launcher": "^2.2.0",
     "karma-firefox-launcher": "^1.1.0",
     "karma-jasmine": "^1.1.2",
-    "protractor": "^5.4.1"
+    "mini-css-extract-plugin": "^0.8.0",
+    "protractor": "^5.4.1",
+    "style-loader": "^1.0.0",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.9.0"
   },
   "scripts": {
-    "postinstall": "npm run copy-libs",
-    "update-deps": "npm update",
-    "postupdate-deps": "npm run copy-libs",
-    "copy-libs": "cpx \"node_modules/{angular,angular-*}/**/*\" app/lib -C",
-    "prestart": "npm install",
-    "start": "http-server -a localhost -p 8000 -c-1 ./app",
-    "pretest": "npm install",
+    "start": "webpack-dev-server --config webpack.config.dev.js --port 3000",
+    "build": "webpack --config webpack.config.dev.js",
     "test": "karma start karma.conf.js",
     "test-single-run": "npm test -- --single-run",
     "preupdate-webdriver": "npm install",
